New Version Created 2.28.2024
This model clips together and gives space for wiring to rest in the base.  Recommended that if you are using a terminal block attach it with wiring holes facing the center of the PMW3360/PMW3389.   Works either way, but the fit is a little tight if you face it out and away from the center of the optical sensor.